requirements.txt
alabaster==0.7.8
appdirs==1.4.3
asgiref==3.5.0
Babel==2.6.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
Click==7.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gpg===1.14.0-unknown
httplib2==0.14.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
itypes==1.2.0
Jinja2==2.10.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.40.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
Sphinx==1.8.5
sqlparse==0.4.2
systemd-python==234
torbrowser-launcher==0.3.2
ttystatus==0.38
ubuntu-advantage-tools==27.2
ufw==0.36
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
xkit==0.0.0
zipp==1.0.0
